Alternative tagging interface
This presents two separate lists, one containing all the topics and subtopics and another with the selected topics.

Topics are selected using a checkbox rather than clicking, this should hopefully make it more obvious when a topic is selected.

The selected subtopics currently don’t show the topic, although this could be added.

A typeahead find feature might also be useful.