Highlight/unhighlight parent correctly
There was a bug which meant that all the parent topics were sometimes unhighlighted when deselecting topics.

This gives the parent topics an ID suffixed with `_parent` and also delimits the topic IDs with underscores so it is easier to find the different parts of the IDs (parent topics and subtopics).

The code now targets the parent topic using the ID meaning there are no more false matches.